requirements.txt
groq
tqdm
uvicorn
fastapi
pydantic
requests
chromadb
python-dotenv
sentence-transformers